% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttf2pt1.r
\docType{package}
\name{Rttf2pt1-package}
\alias{Rttf2pt1}
\alias{Rttf2pt1-package}
\title{Rttf2pt1: 'ttf2pt1' Program}
\description{
Contains the program 'ttf2pt1', for use with the 'extrafont' package. This product includes software developed by the 'TTF2PT1' Project and its contributors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fbertran/Rttf2pt1}
  \item Report bugs at \url{https://github.com/fbertran/Rttf2pt1/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Winston Chang \email{winston@stdout.org}
  \item Andrew Weeks
  \item Frank M. Siegert
  \item Mark Heath
  \item Thomas Henlick
  \item Sergey Babkin
  \item Turgut Uyar
  \item Rihardas Hepas
  \item Szalay Tamas
  \item Johan Vromans
  \item Petr Titera
  \item Lei Wang
  \item Chen Xiangyang
  \item Zvezdan Petkovic
  \item  Rigel
  \item I. Lee Hetherington
}

}
\keyword{internal}
