% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttf2pt1.r
\docType{package}
\name{Rttf2pt1}
\alias{Rttf2pt1}
\alias{package-Rttf2pt1}
\title{Rttf2pt1 package}
\description{
This package is a wrapper for the ttf2pt1 program, and is meant
to be used with the extrafont package.
}
