% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttf2pt1.r
\name{which_ttf2pt1}
\alias{which_ttf2pt1}
\title{Returns the path to the executable for ttf2pt1}
\usage{
which_ttf2pt1()
}
\description{
Returns the path to the executable for ttf2pt1
}
\examples{
which_ttf2pt1()

}
